Fix: refactor

+++ libs/upload_google_drive/__init__.py
@@ -15,37 +15,45 @@
 
 
 from __future__ import print_function
-from datetime import datetime
-import sys
-from typing import Any, Dict, List
-import shutil
 
+import shutil
+import sys
+from concurrent.futures import ThreadPoolExecutor
+from datetime import datetime
+from typing import Any, Dict, List
+
+from tqdm import tqdm
+
+from libs.utils.constants import (CARTOON_DIR, DELETE_FILE_AFTER_UPLOADED,
+                                  DRIVE_CARTOONS_DIR_ID,
+                                  MANGE_EXISTS_FILE_PATH,
+                                  UPDATE_MINUTE_THRESHOLD,
+                                  UPDATE_TIMESTAMP_FILE_PATH)
 from libs.utils.file_helper import mkdir
-from libs.utils.constants import CARTOON_DIR, DRIVE_CARTOONS_DIR_ID, MANGE_EXISTS_FILE_PATH, UPDATE_MINUTE_THRESHOLD, UPDATE_TIMESTAMP_FILE_PATH
 
 sys.path.append("../../libs")  # Adds higher directory to python modules path.
 sys.path.append("../utils")
 
 
-from .interface import ProjectCartoonItem, ProjectItem
-
-from .google_auth import authen
-from googleapiclient.http import MediaFileUpload
-from googleapiclient.errors import HttpError
-from googleapiclient.discovery import build
-from google_auth_oauthlib.flow import InstalledAppFlow
-from google.oauth2.credentials import Credentials
-from google.auth.transport.requests import Request
-import os
 import glob
 # sys.path.append(".")  # Adds higher directory to python modules path.
 import json
+import os
 
+from google.auth.transport.requests import Request
+from google.oauth2.credentials import Credentials
+from google_auth_oauthlib.flow import InstalledAppFlow
+from googleapiclient.discovery import build
+from googleapiclient.errors import HttpError
+from googleapiclient.http import MediaFileUpload
+
+from .google_auth import authen
+from .interface import ProjectCartoonItem, ProjectItem
 
 # If modifying these scopes, delete the file token.json.
 SCOPES = ['https://www.googleapis.com/auth/drive.metadata.readonly']
 
-def upload_to_drive(logging=False):
+def upload_to_drive(logging=False, max_workers=1):
     """
         Upload to drive
     """
@@ -56,11 +64,12 @@ def upload_to_drive(logging=False):
         # drive_project_dirs = list_drive_dirs(service, drive_cartoons_dir_id)
         # Call the Drive v3 API
         # results = service.files().list(
-        #     q = "mimeType = 'application/vnd.google-apps.folder'",
+        #     q = "mimeType = 'application/vnd.google-apps.folder' and trashed=false",
         #     spaces='drive',
         #     pageSize=50, fields="nextPageToken, files(id, name)").execute()
         # items = results.get('files', [])
-
+        
+            
         # project level
         for cartoon_project in cartoon_projects[::]:
             cartoon_project_res = find_or_init_dir(
@@ -82,39 +91,47 @@ def upload_to_drive(logging=False):
                             print(
                             f'\tdir_id: {sub_dir_id}, dir: {sub_dir.get("name")}')
 
+                        manga_files = cartoon_project_sub_dir.image_pdf_list[::]
+
                         # upload manga
-                        for image_pdf in cartoon_project_sub_dir.image_pdf_list[::]:
+                        manga_name = sub_dir.get("name")
+                        for i, image_pdf in tqdm(enumerate(manga_files),
+                                    desc=f'{manga_name}',
+                                    total=len(manga_files)):
                             file_name = image_pdf.split('/')[-1]
                             file_path = image_pdf
-                            if logging:
-                                print(
-                                f'\t\tfile_name: {file_name}, \tfile_path: {file_path}')
-                            upload_manga_res = upload_file_to_drive_if_not_exists(
-                                service, file_name, file_path, sub_dir_id)
 
-                            if upload_manga_res is not None:
-                                file_id, file = upload_manga_res
-                                if logging:
-                                    print(
-                                    f'\t\tfile_id: {file_id}, file: {file.get("name")}')
-                                delete_file(file_path)
+                            upload_file(
+                            service, logging, sub_dir_id, file_name, file_path)
+
 
         # if not drive_project_dirs:
         #     print('No files found.')
         #     return
         # print('Files:')
-        # for project_dir in drive_project_dirs[:1:]:
+        # for project_dir in drive_project_dirs[::]:
         #     print(u'{0} ({1})'.format(project_dir['name'], project_dir['id']))
         #     drive_project_manga_dirs = list_drive_dirs(service, project_dir['id'])
-        #     for manga_dir in drive_project_manga_dirs[:1:]:
+        #     for manga_dir in drive_project_manga_dirs[::]:
         #         print(u'\t{0} ({1})'.format(manga_dir['name'], manga_dir['id']))
         #         drive_manga_chapters = list_drive_manga(service, manga_dir['id'])
-        #         for drive_manga_chapter in drive_manga_chapters[:1:]:
+        #         for drive_manga_chapter in drive_manga_chapters[::]:
         #             print(u'\t\t{0} ({1})'.format(drive_manga_chapter['name'], drive_manga_chapter['id']))
 
     except HttpError as error:
         # TODO(developer) - Handle errors from drive API.
         print(f'An error occurred: {error}')
+
+def upload_file( service, logging: bool, sub_dir_id: str, file_name: str, file_path: str):
+    upload_manga_res = upload_file_to_drive_if_not_exists(
+                                    service, file_name, file_path, sub_dir_id)
+
+    if upload_manga_res is not None:
+        file_id, file = upload_manga_res
+        if logging:
+            print(
+                                        f'\t\tfile_id: {file_id}, file: {file.get("name")}')
+        delete_file(file_path)
 
 
 def generate_drive_manga_exists(force_update = False, logging = False)->Dict[Any, Any]:
@@ -123,8 +140,8 @@ def generate_drive_manga_exists(force_update = False, logging = False)->Dict[Any
     """
     manga_exists_json = {}
     is_early_update = False
-    
-    
+
+
     try:
         if os.path.exists(UPDATE_TIMESTAMP_FILE_PATH):
             with open(UPDATE_TIMESTAMP_FILE_PATH, "r", encoding='utf-8') as read_file:
@@ -135,10 +152,10 @@ def generate_drive_manga_exists(force_update = False, logging = False)->Dict[Any
                     print(f'last update: {timestamp.strftime("%Y-%m-%d %H:%M:%S")}')
                     time_diff = now - timestamp
                     if time_diff.total_seconds() / 60 < UPDATE_MINUTE_THRESHOLD:
-                        
+
                         is_early_update = True
-            
-        
+
+
         # Opening JSON file
         if os.path.exists(MANGE_EXISTS_FILE_PATH):
             with open(MANGE_EXISTS_FILE_PATH, "r", encoding='utf-8') as read_file:
@@ -171,12 +188,12 @@ def generate_drive_manga_exists(force_update = False, logging = False)->Dict[Any
                 if logging:
                     print(u'\t{0} ({1})'.format(
                     manga_dir['name'], manga_dir['id']))
-                
+
                 manga_exists_json[project_dir['name']]["sub_dirs"][manga_dir['name']] = {
                     "id": manga_dir['id'],
                     "chapters": {}
                 }
-                
+
                 drive_manga_chapters = list_drive_manga(
                     service, manga_dir['id'])
                 for drive_manga_chapter in drive_manga_chapters[::]:
@@ -185,15 +202,21 @@ def generate_drive_manga_exists(force_update = False, logging = False)->Dict[Any
                         drive_manga_chapter['name'], drive_manga_chapter['id']))
                     manga_exists_json[project_dir['name']]["sub_dirs"][manga_dir['name']]["chapters"][drive_manga_chapter['name']] = {
                     "id": drive_manga_chapter['id'],
+                    "createdTime": drive_manga_chapter['createdTime'], 
+                    "modifiedByMeTime": drive_manga_chapter['modifiedByMeTime']
                 }
-                    
+
+                manga_exists_json[project_dir['name']]["sub_dirs"][manga_dir['name']]["chapters"] = orderKeysInJson(manga_exists_json[project_dir['name']]["sub_dirs"][manga_dir['name']]["chapters"])
+            manga_exists_json[project_dir['name']]["sub_dirs"] = orderKeysInJson(manga_exists_json[project_dir['name']]["sub_dirs"])   
+        manga_exists_json = orderKeysInJson(manga_exists_json)
+
         # Serializing json
         json_object = json.dumps(manga_exists_json, indent=2, ensure_ascii=False)
-        
+
         # Writing to sample.json
         with open(MANGE_EXISTS_FILE_PATH, "w", encoding='utf-8') as outfile:
-            outfile.write(json_object)    
-        
+            outfile.write(json_object)
+
         with open(UPDATE_TIMESTAMP_FILE_PATH, "w", encoding='utf-8') as outfile:
             # Getting the current date and time
             dt = datetime.now()
@@ -226,15 +249,12 @@ def list_drive_dirs(service, drive_cartoons_dir_id: str):
     while True:
         # pylint: disable=maybe-no-member
         response = service.files().list(
-            q=f"'{drive_cartoons_dir_id}' in parents",
-            # q="'1iXUAF2N3YxLPvJyDTYYL1f8HFP1mU1Ek' in parents,mimeType='application/vnd.google-apps.folder'",
+            q=f"'{drive_cartoons_dir_id}' in parents and trashed=false",
             spaces='drive',
             fields='nextPageToken, '
             'files(id, name)',
             pageToken=page_token).execute()
-        # for file in response.get('files', []):
-        #     # Process change
-        #     print(F'Found file: {file.get("name")}, {file.get("id")}')
+
         drive_project_dirs.extend(response.get('files', []))
         page_token = response.get('nextPageToken', None)
         if page_token is None:
@@ -248,15 +268,12 @@ def list_drive_manga(service, drive_cartoons_dir_id: str):
     while True:
         # pylint: disable=maybe-no-member
         response = service.files().list(
-            q=f"'{drive_cartoons_dir_id}' in parents",
-            # q="'1iXUAF2N3YxLPvJyDTYYL1f8HFP1mU1Ek' in parents,mimeType='application/vnd.google-apps.folder'",
+            q=f"'{drive_cartoons_dir_id}' in parents and trashed=false",
             spaces='drive',
             fields='nextPageToken, '
-            'files(id, name)',
+            'files(id, name, createdTime, modifiedByMeTime)',
             pageToken=page_token).execute()
-        # for file in response.get('files', []):
-        #     # Process change
-        #     print(F'Found file: {file.get("name")}, {file.get("id")}')
+
         drive_project_dirs.extend(response.get('files', []))
         page_token = response.get('nextPageToken', None)
         if page_token is None:
@@ -266,21 +283,21 @@ def list_drive_manga(service, drive_cartoons_dir_id: str):
 
 def upload_file_to_drive_if_not_exists(service, file_name: str, file_path: str, folder_id: str):
     try:
-        dir = None
+        dir_result = None
         dir_res = service.files().list(
-            q=f"'{folder_id}' in parents and name='{file_name}' and mimeType='application/pdf'",
+            q=f"'{folder_id}' in parents and name='{file_name}' and mimeType='application/pdf' and trashed=false",
             spaces='drive',
             fields='nextPageToken, '
-            'files(id, name)',
+            'files(id, name, createdTime, modifiedByMeTime)',
         ).execute()
         # print(dir_res)
         if dir_res is not None and dir_res['files'] is not None and len(dir_res['files']) > 0:
-            dir = dir_res['files'][0]
+            dir_result = dir_res['files'][0]
         # print(f'dir_name: {dir_name}')
-        # print(dir)
-        if dir == None:
+        # print(dir_result)
+        if dir_result == None:
             return upload_file_to_drive(service, file_name, file_path, folder_id)
-        return dir.get('id'), dir
+        return dir_result.get('id'), dir_result
     except Exception as e:
         print(e)
         return None
@@ -299,17 +316,17 @@ def upload_file_to_drive(service, file_name: str, file_path: str, folder_id: str
                                   media_body=media,
                                   fields='id, name').execute()
     print(file)
-    print(u'Upload File ID: %s'.format(file.get('id')))
-    return file.get('id'), file
+    return file['id'], file
 
 
 def get_cartoon_projects() -> List[ProjectItem]:
+    """ get_cartoon_projects """
     projects: List[ProjectItem] = []
     for project_dir in os.listdir(f'{CARTOON_DIR}'):
         # for name in files:
         #     print(os.path.join(root, name))
         project_dir_path = os.path.join(CARTOON_DIR, project_dir)
-        if project_dir == None or not os.path.isdir(project_dir_path):
+        if project_dir is None or not os.path.isdir(project_dir_path):
             continue
         title = project_dir
 
@@ -317,7 +334,7 @@ def get_cartoon_projects() -> List[ProjectItem]:
 
         for sub_project_cartoon_dir in os.listdir(project_dir_path):
             # print(f'{sub_project_cartoon_dir}')
-            if sub_project_cartoon_dir == None:
+            if sub_project_cartoon_dir is None:
                 continue
             image_pdf_list = glob.glob(os.path.join(
                 CARTOON_DIR, project_dir, sub_project_cartoon_dir, '*.pdf'))
@@ -329,22 +346,23 @@ def get_cartoon_projects() -> List[ProjectItem]:
 
 
 def find_or_init_dir(service, drive_cartoons_dir_id: str, dir_name: str):
-    dir = None
+    """ find_or_init_dir """
+    dir_file = None
     dir_res = service.files().list(
-        q=f"'{drive_cartoons_dir_id}' in parents and name='{dir_name}' and mimeType='application/vnd.google-apps.folder'",
+        q=f"'{drive_cartoons_dir_id}' in parents and name='{dir_name}' and mimeType='application/vnd.google-apps.folder' and trashed=false",
         spaces='drive',
         fields='nextPageToken, '
-        'files(id, name)',
+        'files(*)',
     ).execute()
-    # print(dir_res)
+    print(f'dir_res: {dir_res}')
     if dir_res is not None and dir_res['files'] is not None and len(dir_res['files']) > 0:
-        dir = dir_res['files'][0]
+        dir_file = dir_res['files'][0]
     # print(dir)
     # print(f'dir_name: {dir_name}')
     # print(dir)
-    if dir == None:
+    if dir_file is None:
         return create_folder(service, drive_cartoons_dir_id, dir_name)
-    return dir.get('id'), dir
+    return dir_file.get('id'), dir_file
 
 
 def create_folder(service, parent_dir_id: str, dir_name: str):
@@ -380,10 +398,16 @@ def delete_file(file_path: str):
         Delete file after upload to drive
     """
     if os.path.exists(file_path):
-        # trash_dir = os.path.join(
-        #     'files', 'trash', '/'.join(file_path.split('/')[1:-1:]))
-        # print(f'trash_dir: {trash_dir}')
-        # mkdir(trash_dir)
-        # new_path = os.path.join(trash_dir, file_path.split('/')[-1])
-        # shutil.move(file_path, new_path)
-        os.remove(file_path)
+        if DELETE_FILE_AFTER_UPLOADED:
+            os.remove(file_path)
+        else:
+            trash_dir = os.path.join(
+                'files', 'trash', '/'.join(file_path.split('/')[2:-1:]))
+            # print(f'trash_dir: {trash_dir}')
+            mkdir(trash_dir)
+            new_path = os.path.join(trash_dir, file_path.split('/')[-1])
+            shutil.move(file_path, new_path)
+
+
+def orderKeysInJson(data: dict)->dict:
+    return dict(sorted(data.items(), key=lambda x: str(x[0]).zfill(30)))
